backend/ai-router/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
qdrant-client
requests
PyPDF2
python-docx
python-multipart
